[IMP] html_builder: savePlugin should be optional
Make the `savePlugin` optional as it won't be used in `mass_mailing` (saving is
done through the Form view).

Save/discard handling is mostly delegated to the `website_builder`.

Add a slot to insert custom buttons in the top-actions toolbar of the
html_builder (mass_mailing will use `toggleCodeView` and `toggleFullScreen`, and
won't use `save` nor `discard` as these actions are already handled by the form
view.

task-4247642
